refactor(logic): NodeParser를 interface 없도록, Operator.of() 함수 추가
