+++ src/test/kotlin/calculator/parser/NodeParserTest.kt
@@ -6,10 +6,10 @@ import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
 
-class SimpleNodeParserTest {
+class NodeParserTest {
     @Test
     fun `parse() 예제 입력 테스트`() {
-        assertThat(SimpleNodeParser().parse("2 + 3 * 4 / 2"))
+        assertThat(NodeParser().parse("2 + 3 * 4 / 2"))
             .isEqualTo(
                 listOf(
                     Number(2.0),
@@ -25,7 +25,7 @@ class SimpleNodeParserTest {
 
     @Test
     fun `parse() 다양한 입력 검증`() {
-        assertThat(SimpleNodeParser().parse("123456 +++   0 *- 4 4 444 // 2    "))
+        assertThat(NodeParser().parse("123456 +++   0 *- 4 4 444 // 2    "))
             .isEqualTo(
                 listOf(
                     Number(123456.0),
@@ -48,14 +48,14 @@ class SimpleNodeParserTest {
     @Test
     fun `parse() 잘못된 입력이 있는 경우 InvalidCharacter 발생`() {
         assertThatThrownBy {
-            assertThat(SimpleNodeParser().parse("123 * 12 _ 3"))
-        }.isInstanceOf(NodeParser.Error.InvalidCharacter::class.java)
+            assertThat(NodeParser().parse("123 * 12 _ 3"))
+        }.isInstanceOf(Operator.InvalidCharacter::class.java)
             .hasMessage("Invalid character _ has found in text.")
     }
 
     @Test
     fun `trimAll() 각 리스트의 모든 문자열 trim 테스트`() {
-        SimpleNodeParser().apply {
+        NodeParser().apply {
             assertThat(listOf("hello ", "kotlin! ", " ", "world").trimAll())
                 .isEqualTo(listOf("hello", "kotlin!", "", "world"))
         }
@@ -63,7 +63,7 @@ class SimpleNodeParserTest {
 
     @Test
     fun `exceptEmpty() 공백 문자열은 리스트에서 제외 테스트`() {
-        SimpleNodeParser().apply {
+        NodeParser().apply {
             assertThat(listOf("", " ", "! ", "", "", "hello").exceptEmpty())
                 .isEqualTo(listOf(" ", "! ", "hello"))
         }
@@ -71,7 +71,7 @@ class SimpleNodeParserTest {
 
     @Test
     fun `convertToNode() 문자열을 Node로 변환`() {
-        SimpleNodeParser().apply {
+        NodeParser().apply {
             assertThat("123".convertToNode()).isEqualTo(Number(123.0))
             assertThat("0".convertToNode()).isEqualTo(Number(0.0))
             assertThat("123".convertToNode()).isEqualTo(Number(123.0))
@@ -85,23 +85,23 @@ class SimpleNodeParserTest {
     @Test
     fun `convertToNode() 잘못된 문자열 InvalidCharacter 반환`() {
         assertThatThrownBy {
-            SimpleNodeParser().apply {
+            NodeParser().apply {
                 "XXX".convertToNode()
             }
-        }.isInstanceOf(NodeParser.Error.InvalidCharacter::class.java)
+        }.isInstanceOf(Operator.InvalidCharacter::class.java)
             .hasMessage("Invalid character XXX has found in text.")
 
         assertThatThrownBy {
-            SimpleNodeParser().apply {
+            NodeParser().apply {
                 "_".convertToNode()
             }
-        }.isInstanceOf(NodeParser.Error.InvalidCharacter::class.java)
+        }.isInstanceOf(Operator.InvalidCharacter::class.java)
             .hasMessage("Invalid character _ has found in text.")
     }
 
     @Test
     fun `convertAllToNode() 문자열 리스트를 Node 리스트로 변환`() {
-        SimpleNodeParser().apply {
+        NodeParser().apply {
             assertThat(
                 listOf("123", "0", "123", "+", "-", "*", "/").convertAllToNode()
             ).isEqualTo(
